fix user storage issue and read vnoc.conf correctly

+++ NetService/src/FileUserStorage.h
@@ -11,7 +11,7 @@ public:
 	~fUserStorage();
 		//返回密码的长度
 	bool IfUserExist(const char* pUser);
-	bool GetPassword(const char* pUser, char* pPassWordBuff, long cbBuff);
+	bool GetPassword(const char* pUser, char* pPassWordBuff, size_t cbBuff);
 	bool GetUserInfo(const char* pUser, userinfo* pUserInfo);
 
 private:
