+++ NetService/src/UserStorage.h
@@ -10,7 +10,7 @@ public:
 	UserStorage();
 	~UserStorage();
 	virtual bool IfUserExist(const char* pUser);
-	virtual bool GetPassword(const char* pUser, char* pPassWordBuff, long cbBuff) ;
+	virtual bool GetPassword(const char* pUser, char* pPassWordBuff, size_t cbBuff) ;
 	virtual bool GetUserInfo(const char* pUser, userinfo* pUserInfo) ;
 private:
